% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc_codes.R
\docType{data}
\name{atc_codes}
\alias{atc_codes}
\title{Hierarchical tree of the ATC system for classifying drugs}
\usage{
data(atc_codes)
}
\description{
A dataset including the following column:
\describe{
\item{\code{pathString}}{A string identifying all the parents of a node. Each
parent is separated by a \code{/}.}
}
}
\keyword{datasets}
