% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Creating a tree file for further use in TreeScan}
\usage{
create_tree(x)
}
\arguments{
\item{x}{A data frame that includes two columns:
\itemize{
  \item{node}
  \item{parent}
  }}
}
\description{
Creating a tree file for further use in TreeScan
}
