% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses.R
\docType{data}
\name{diagnoses}
\alias{diagnoses}
\title{Test dataset of ICD diagnoses}
\format{
A data frame with 23,144 rows and 3 columns
}
\usage{
data(diagnoses)
}
\description{
A simulated dataset of hospital diagnoses created with the help of
the \code{comorbidity} package including the following columns:
\describe{
\item{id}{Individual identifier,}
\item{case}{Indicator for case status,}
\item{diag}{An ICD-10 diagnosis code.}
}
}
\keyword{datasets}
