% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_10_se.R
\docType{data}
\name{icd_10_se}
\alias{icd_10_se}
\title{Swedish version of the ICD-10 diagnoses code tree}
\usage{
data(icd_10_se)
}
\description{
A dataset including the following column:
\describe{
\item{\code{pathString}}{A string identifying all the parents of a node. Each
parent is separated by a \code{/}.}
}
}
\keyword{datasets}
