% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_10_se_dict.R
\docType{data}
\name{icd_10_se_dict}
\alias{icd_10_se_dict}
\title{Dictionary for the Swedish version of the ICD-10 diagnoses code tree}
\usage{
data(icd_10_se_dict)
}
\description{
A dataset including the following column:
\describe{
   \item{`node`}{A string identidying a node}
   \item{`title`}{A label for the node}
}
}
\keyword{datasets}
