% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMineR.R
\name{TreeMineR}
\alias{TreeMineR}
\title{Tree scan statistics for R}
\usage{
TreeMineR(
  data,
  tree,
  dictionary = NULL,
  delimiter = "/",
  p = NULL,
  n_monte_carlo_sim = 9999,
  random_seed = FALSE,
  future_control = list(strategy = "sequential")
)
}
\arguments{
\item{data}{The dataset used for the computation. The dataset needs to include the
   following columns:

   \describe{
   \item{`id`}{An integer that is unique to every individual.}
   \item{`leaf`}{A string identifying the unique diagnoses or leafs for each individual.}
   \item{`exposed`}{A 0/1 indicator of the individual's exposure status.}
   }
   See below for the first and last rows included in the example dataset.

   ```
      id leaf exposed
       1 K251       0
       2 Q702       0
       3  G96       0
       3 S949       0
       4 S951       0
    ---
     999 V539       1
     999 V625       1
     999 G823       1
    1000  L42       1
    1000 T524       1
   ```}

\item{tree}{A dataset with one variable `pathString` defining the tree structure
that you would like to use. This dataset can, e.g., be created using
 \code{\link{create_tree}}.}

\item{dictionary}{A data.frame that includes one `node` column and a `title` column,
which are used for labeling the cuts in the output of `TreeMineR`.}

\item{delimiter}{A character defining the delimiter of different tree levels within your
`pathString`. The default is `/`.}

\item{p}{The proportion of exposed individuals in the dataset. The default is the
proportion of exposed individuals among unique individuals in `data`}

\item{n_monte_carlo_sim}{The number of Monte-Carlo simulations to be used for calculating P-values.}

\item{random_seed}{Random seed used for the Monte-Carlo simulations.}

\item{future_control}{A list of arguments passed `future::plan`. This is useful if one would like
to parallelise the Monte-Carlo simulations to decrease the computation
time. The default is a sequential run of the Monte-Carlo simulations.}
}
\description{
Tree scan statistics for R
}
\examples{
TreeMineR(data = diagnoses,
          tree  = icd_10_se,
          p = 1/11,
          n_monte_carlo_sim = 99,
          random_seed = 1234)

}
