% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeScan.R
\name{TreeScan}
\alias{TreeScan}
\title{Tree scan statistics for R}
\usage{
TreeScan(
  data,
  exposure,
  p = FALSE,
  leafs,
  id,
  n_monte_carlo_sim = 1000,
  random_seed = FALSE,
  n_level = FALSE,
  future_control = list(strategy = "sequential")
)
}
\arguments{
\item{data}{The dataset used for the computation}

\item{exposure}{The name of the exposure variable}

\item{p}{The proportion of exposed individuls in the dataset. The default is the
proportion of exposed individuals among unique individuals in `data`.}

\item{leafs}{The name of the leaf variable}

\item{id}{The name of the id variable}

\item{n_monte_carlo_sim}{The number of Monte Carolo simulations used for calculating P-values.}

\item{n_level}{The number of levels of the tree. The default is the maximum character
length found among the leafs.}

\item{parallel}{Logical indicator for whether the Monte Carlo simulations should be
parallelised.}

\item{parallel_control}{A list of arguments passed to `makeCluster` for parallisation.}
}
\description{
Tree scan statistics for R
}
\examples{
TreeScan(diagnoses,
         exposure = case,
         leafs = diag,
         id = id,
         n_monte_carlo_sim = 10,
         parallel = TRUE,
         parallel_control = list(2))

}
