% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeScan.R
\name{TreeScan}
\alias{TreeScan}
\title{Tree scan statistics for R}
\usage{
TreeScan(
  counts,
  tree,
  delimiter = "/",
  p,
  n_monte_carlo_sim = 9999,
  random_seed = FALSE,
  future_control = list(strategy = "sequential")
)
}
\arguments{
\item{counts}{A data set including the number of events among exposed and unexposed for
 each lead. The dataset needs to include the following columns:
 \itemize{
   \item{`leaf`}{The code defining the leaf.}
   \item{`n0`}{The number of events within the leaf among the unexposed.}
   \item{`n1`}{The number of events within the leaf among the exposed.}
   }

   See below for the first and last rows included in the example dataset.

   ```
   leaf n0 n1
   1: K251  1  0
   2: Q702  5  0
   3:  G96  3  0
   4: S949  2  0
   5: S951  2  0
   ---
   9492: T118  0  1
   9493: A932  0  1
   9494: D350  0  1
   9495: L410  0  1
   9496: T524  0  1
   ```}

\item{tree}{A dataset with one variable `pathString` defining the tree structure
that you would like to use. This dataset can, e.g., be created using
 \code{\link{create_tree}}.}

\item{delimiter}{A character defining the delimiter of different tree levels within your
`pathString`. The default is `/`.}

\item{p}{The proportion of exposed individuals in the dataset. The default is the
proportion of exposed individuals among unique individuals in `data`.}

\item{n_monte_carlo_sim}{The number of Monte-Carlo simulations to be used for calculating P-values.}

\item{random_seed}{Random seed used for the Monte-Carlo simulations.}

\item{future_control}{A list of arguments passed `future::plan`. This is useful if one would like
to parallelise the Monte-Carlo simulations to decrease the computation
time. The default is a sequential run of the Monte-Carlo simulations.}
}
\description{
Tree scan statistics for R
}
\examples{
TreeScan(diagnoses,
         exposure = case,
         leafs = diag,
         id = id,
         n_monte_carlo_sim = 10)

}
